created interface between ezmesh and gmsh
