+++ ezmesh/__init__.py
@@ -1,1 +1,3 @@
-from ezmesh.mesh import *
+from ezmesh.geometry import Geometry, BoundaryLayer, CurveLoop, PlaneSurface, Point, Line
+from ezmesh.mesh import Mesh, ElementType
+from ezmesh.visualizer import visualize_mesh
