+++ requirements_dev.txt
@@ -5,3 +5,4 @@ matplotlib
 
 numpy
 gmsh
+pythreejs
